requirements.txt
alembic==1.6.5
aniso8601==9.0.1
apispec==4.5.0
attrs==21.2.0
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
psycopg2-binary==2.8.6
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==2.0.1
